refactor: split sleeper_tools — extract transactions (final domain split)
Move get_transactions (week-inferring, robust snapshot fallback) and get_traded_picks into nfl_mcp/sleeper_transactions.py. It imports the primitives it consumes from core (get_nfl_state, _init_db, _enrich_single) + enrichment (_enrich_usage_and_opponent); sleeper_tools re-exports the two at the end of the module (after get_nfl_state) to stay acyclic. sleeper_tools.py: 1372 -> 1108 lines (3022 -> 1108 across all splits, -63%). Added the B023 per-file-ignore for sleeper_transactions.py (the same-iteration enrichment closures moved with get_transactions). Repointed 3 transactions white-box tests to patch sleeper_transactions (create_http_client / get_nfl_state). ruff clean; full suite 902 passed, 2 skipped. With enrichment/strategy/transactions extracted, the reviewer's league/roster/transactions/enrichment decomposition is complete (league/roster = the core); matchups/drafts remain as optional further extractions.
